Changing the project's structure so the pipeline could pass
