+++ where_to_go/lib/my_app.dart
@@ -1,0 +1,11 @@
+import "package:flutter/material.dart";
+import "home_screen.dart";
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const HomeScreen();
+  }
+}
